requirements.txt
aws-cdk-lib==2.147.2
constructs>=10.0.0,<11.0.0

boto3==1.28.78
botocore==1.31.78
